+++ src/components/guidespage/SubLinks.tsx
@@ -10,6 +10,7 @@ type SubLinksProps = {
 
 const SubLinks: FC<SubLinksProps> = ({ fileList, onLinkClick }) => {
   const { activeLink, setActiveLink } = useDocumentationContext();
+  console.log(activeLink);
 
   fileList = [
     ...fileList.filter((file) => file.name === 'welcome'),
